Renamed AboutPageSectionComponent
